♻️ Delete file_access and move content to dev.storage
